--- a/src/optimization/environment/index.ts
+++ b/src/optimization/environment/index.ts
@@ -1,0 +1,8 @@
+// Environment detection and platform-specific strategies
+// Week 2 deliverables - Agent B
+
+export * from './EnvironmentDetector';
+export * from './RequestStrategy';
+export * from './strategies';
+
+// TODO(Agent B): Implement during Week 2