--- a/src/optimization/environment/strategies/index.ts
+++ b/src/optimization/environment/strategies/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Environment Strategy Exports
+ */
+
+export { RequestStrategyFactory } from '../RequestStrategy';
+export { EnvironmentDetector } from '../EnvironmentDetector';
+
+// Re-export strategy types
+export type { RequestStrategy } from '../RequestStrategy';