… .claude.backup-2025-08-04T20-47-55-221Z/agents/bug-root-cause-analyzer.md
… .claude.backup-2025-08-04T20-47-55-221Z/agents/spec-breaking-change-detector.md
… .claude.backup-2025-08-04T20-47-55-221Z/agents/spec-completion-reviewer.md
… .claude.backup-2025-08-04T20-47-55-221Z/agents/spec-dependency-analyzer.md
… .claude.backup-2025-08-04T20-47-55-221Z/agents/spec-design-validator.md
… .claude.backup-2025-08-04T20-47-55-221Z/agents/spec-design-web-researcher.md
… .claude.backup-2025-08-04T20-47-55-221Z/agents/spec-documentation-generator.md
… .claude.backup-2025-08-04T20-47-55-221Z/agents/spec-duplication-detector.md
… .claude.backup-2025-08-04T20-47-55-221Z/agents/spec-integration-tester.md
… .claude.backup-2025-08-04T20-47-55-221Z/agents/spec-performance-analyzer.md
… .claude.backup-2025-08-04T20-47-55-221Z/agents/spec-requirements-validator.md
… .claude.backup-2025-08-04T20-47-55-221Z/agents/spec-task-executor.md
… .claude.backup-2025-08-04T20-47-55-221Z/agents/spec-task-implementation-reviewer.md
… .claude.backup-2025-08-04T20-47-55-221Z/agents/spec-task-validator.md
… .claude.backup-2025-08-04T20-47-55-221Z/agents/spec-test-generator.md
… .claude.backup-2025-08-04T20-47-55-221Z/agents/steering-document-updater.md
… .claude.backup-2025-08-04T20-47-55-221Z/bugs/agent-enablement/analysis.md
… .claude.backup-2025-08-04T20-47-55-221Z/bugs/agent-enablement/report.md
… .claude.backup-2025-08-04T20-47-55-221Z/bugs/agent-enablement/verification.md
… .claude.backup-2025-08-04T20-47-55-221Z/commands/bug-analyze.md
… .claude.backup-2025-08-04T20-47-55-221Z/commands/bug-create.md
… .claude.backup-2025-08-04T20-47-55-221Z/commands/bug-fix.md
… .claude.backup-2025-08-04T20-47-55-221Z/commands/bug-status.md
… .claude.backup-2025-08-04T20-47-55-221Z/commands/bug-verify.md
… .claude.backup-2025-08-04T20-47-55-221Z/commands/simplify-git-installation/task-17.md
… .claude.backup-2025-08-04T20-47-55-221Z/commands/simplify-git-installation/task-18.md
… .claude.backup-2025-08-04T20-47-55-221Z/commands/simplify-git-installation/task-19.md
… .claude.backup-2025-08-04T20-47-55-221Z/commands/simplify-git-installation/task-20.md
… .claude.backup-2025-08-04T20-47-55-221Z/commands/simplify-git-installation/task-21.md
… .claude.backup-2025-08-04T20-47-55-221Z/commands/simplify-git-installation/task-22.md